Add a workaround for removing the column in the decidim_attachments table

+++ decidim-core/db/migrate/20251112132305_remove_legacy_images_from_core_module.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
  
-class RemoveLegacyImagesFromCoreModule< ActiveRecord::Migration[7.2]
+class RemoveLegacyImagesFromCoreModule < ActiveRecord::Migration[7.2]
   def change
     remove_column :decidim_organizations, :logo, :string
     remove_column :decidim_organizations, :official_img_footer, :string
@@ -10,7 +10,12 @@ class RemoveLegacyImagesFromCoreModule< ActiveRecord::Migration[7.2]
 
     remove_column :decidim_authorizations, :verification_attachment, :string
 
-    remove_column :decidim_attachments, :file, :string
+    # The original decidim_attachments table creation was in decidim-participatory_processes
+    # at decidim-participatory_processes/db/migrate/20161116115156_create_attachments.rb
+    #
+    # We need to workaround this issue as when creating new application this table may not exist
+    # when this migration is run
+    remove_column :decidim_attachments, :file, :string if table_exists?(:decidim_attachments)
 
     remove_column :decidim_users, :avatar, :string
